Restructure assets/js dir
Broke it out into hooks and core dirs. I've been meaning to do this for
a while.
