+++ assets/js/dreadnought/main.js
@@ -1,13 +1,11 @@
-import CommandCardHooks from "./command_card.js";
-import SliderHooks from "./constant_game_time.js";
-import UnitHooks from "./unit.js";
-import { CheckGsapLoad } from "./gsap_experiment.js";
-import SegmentHooks from "./segment.js";
+import CommandCardHooks from "./hooks/command_card.js";
+// import SliderHooks from "./constant_game_time.js";
+import UnitHooks from "./hooks/unit.js";
+// import { CheckGsapLoad } from "./gsap_experiment.js";
+import SegmentHooks from "./hooks/segment.js";
 
 export default {
   ...CommandCardHooks,
-  ...SliderHooks, 
   ...UnitHooks, 
   ...SegmentHooks,
-  CheckGsapLoad,
 }
